requirements.txt
Flask==2.1.1
Flask_Cors==3.0.10
pylsl==1.15.0
pyserial==3.5
PyYAML==6.0
setuptools==61.2.0
